# docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  model.AddToCartInput:
    properties:
      item_id:
        type: integer
      quantity:
        type: integer
    type: object
...
  model.Brand:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Buyer:
    properties:
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      id:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
    type: object
...
  model.CartItem:
    properties:
      buyer:
        $ref: '#/definitions/model.Buyer'
      buyer_id:
        type: string
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      quantity:
        type: integer
    type: object
  model.Category:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
  model.CategoryInput:
    properties:
      name:
        type: string
    type: object
  model.CategoryOutput:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.Image:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      url:
        type: string
    type: object
  model.Item:
    properties:
      article:
        type: string
      brand:
        $ref: '#/definitions/model.Brand'
      brand_id:
        type: integer
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: integer
      description:
        type: string
      favoritedBy:
        items:
          $ref: '#/definitions/model.Buyer'
        type: array
      height:
        type: integer
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/model.Image'
        type: array
      length:
        type: integer
      material:
        $ref: '#/definitions/model.Material'
      material_id:
        type: integer
      name:
        type: string
      price:
        type: number
      price_with_discount:
        type: number
      quantity:
        type: integer
      seller:
        $ref: '#/definitions/model.Seller'
      seller_id:
        type: string
      weight:
        type: integer
      width:
        type: integer
    type: object
...
  model.Material:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
    type: object
...
  model.Order:
    properties:
      buyer:
        $ref: '#/definitions/model.Buyer'
      buyer_id:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      status:
        type: string
      total:
        type: number
    type: object
  model.OrderItem:
    properties:
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      item_id:
        type: integer
      order:
        $ref: '#/definitions/model.Order'
      order_id:
        type: integer
      quantity:
        type: integer
      total:
        type: number
      unit_price:
        type: number
    type: object
  model.Seller:
    properties:
      balance:
        type: number
      email:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      name:
        type: string
      password:
        type: string
      shop_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/category:
    delete:
      description: Delete a category by ID
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID
        in: query
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            type: string
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to delete category
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a category
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new category in the system
      parameters:
      - description: Bearer {JWT}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Category created successfully
          schema:
            type: string
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: You are not authorized to access this resource
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create category
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new category
      tags:
      - Categories
...
  /cart:
    get:
      description: Retrieves the contents of the buyer's cart. Only accessible by
        buyers.
      produces:
      - application/json
      responses:
        "200":
          description: Cart items
          schema:
            items:
              $ref: '#/definitions/model.CartItem'
            type: array
        "403":
          description: You are not authorized to access this resource
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get cart contents
      tags:
      - cart
  /cart/add:
    post:
      consumes:
      - application/json
      description: Adds a specified item with a specified quantity to the buyer's
        cart. Only accessible by buyers.
      parameters:
      - description: Item ID and Quantity
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.AddToCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: Item added to cart
          schema:
            type: string
        "400":
          description: Invalid input
          schema:
            type: string
        "403":
          description: You are not authorized to access this resource
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Add an item to cart
      tags:
      - cart
  /cart/remove:
    delete:
      description: Removes a specified item from the buyer's cart by cart item ID.
        Only accessible by buyers.
      parameters:
      - description: Cart Item ID
        in: query
        name: cart_item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Item removed from cart
          schema:
            type: string
        "400":
          description: Invalid cart item ID
          schema:
            type: string
        "403":
          description: You are not authorized to access this resource
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Remove an item from the cart
      tags:
      - cart
  /category:
    get:
      description: Retrieve a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            items:
              $ref: '#/definitions/model.CategoryOutput'
            type: array
        "500":
          description: Failed to retrieve category list
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get category list
      tags:
      - Categories
...
swagger: "2.0"